Add wallet policy

+++ app/Http/Controllers/WalletController.php
@@ -5,11 +5,15 @@ namespace App\Http\Controllers;
 use App\Models\Wallet;
 use App\Http\Requests\StoreWalletRequest;
 use App\Http\Requests\UpdateWalletRequest;
+use Illuminate\Support\Facades\Gate;
 
 class WalletController extends Controller
 {
     public function index()
     {
+        // Gate
+        Gate::authorize('viewAny', Wallet::class);
+
         $data = Wallet::all();
 
         return response()->json([
@@ -20,6 +24,9 @@ class WalletController extends Controller
 
     public function store(StoreWalletRequest $request)
     {
+        // Gate
+        Gate::authorize('create', Wallet::class);
+
         $newWallet = Wallet::create($request->validated());
 
         return response()->json([
@@ -31,6 +38,9 @@ class WalletController extends Controller
 
     public function show(Wallet $wallet)
     {
+        // Gate
+        Gate::authorize('view', Wallet::class);
+
         return response()->json([
             "data" => $wallet
         ]);
@@ -39,11 +49,14 @@ class WalletController extends Controller
 
     public function update(UpdateWalletRequest $request, Wallet $wallet)
     {
+        // Gate
+        Gate::authorize('update', Wallet::class);
+
         // Check name exist
         $exist = Wallet::where('name', $request->validated('name'))
             ->where('id', "<>", $wallet->id)
             ->exists();
-            
+
         if ($exist) abort(422, "Bu hisob allaqachon mavjud");
 
         $wallet->update($request->validated());
@@ -57,6 +70,9 @@ class WalletController extends Controller
 
     public function destroy(Wallet $wallet)
     {
+        // Gate
+        Gate::authorize('delete', Wallet::class);
+
         $wallet->delete();
 
         return response()->json([
